+++ src/components/Instructions/InstructionItem.tsx
@@ -2,7 +2,7 @@ import { argType, mapInstructionsArgsByType, valueToBinary, valueToNumeralSystem
 import classNames from "classnames";
 import { ExpectedState, RegistersArray, Status } from "@/types/pvm";
 import { InstructionMode } from "./types";
-import { ForwardedRef, forwardRef, useCallback, useContext, useState } from "react";
+import { ForwardedRef, forwardRef, useCallback, useContext, useMemo, useState } from "react";
 import { NumeralSystemContext } from "@/context/NumeralSystemContext";
 import { TableCell, TableRow } from "../ui/table";
 import { ProgramRow } from "./InstructionsTable";
@@ -13,6 +13,7 @@ import { Tooltip, TooltipContent, TooltipPortal, TooltipTrigger } from "@/compon
 import { useIsDarkMode } from "@/packages/ui-kit/DarkMode/utils";
 import { selectProgram } from "@/store/debugger/debuggerSlice.ts";
 import { getStatusColor } from "@/utils/colors";
+import { NumeralSystem } from "@/context/NumeralSystem";
 
 const getWorkerValueFromState = (
   worker: WorkerState,
@@ -166,6 +167,7 @@ export const InstructionItem = forwardRef(
               <TableCell className="p-1.5 whitespace-nowrap border-b font-inconsolata">
                 <span className="">
                   {"args" in programRow && (
+                    /** Width of this cell is adjusted to the widest item by WidestInstructionItem component */
                     <span
                       dangerouslySetInnerHTML={{
                         __html:
@@ -193,7 +195,6 @@ export const InstructionItem = forwardRef(
               bits={valueToBinary(programRow.instructionCode, 8)}
               value={valueToNumeralSystem(programRow.instructionCode, numeralSystem)}
             />
-
             {"args" in programRow &&
               mapInstructionsArgsByType(programRow.args, numeralSystem, programRow.counter, program)
                 ?.filter((instruction) => !instruction.hiddenFromDetails)
@@ -327,5 +328,61 @@ function DetailsColumn({ kind, bits, value }: DetailsColumnProps) {
         </div>
       </div>
     </div>
+  );
+}
+
+/** A hiddent item of instruction table that ensures the table is adjusted to the widest instruction item */
+export function WidestInstructionItem({
+  instructionMode,
+  programRows,
+}: {
+  instructionMode: InstructionMode;
+  programRows: ProgramRow[] | undefined;
+}) {
+  const program = useAppSelector(selectProgram);
+
+  const widestItem = useMemo(
+    () =>
+      programRows?.reduce<{ widestItem: ProgramRow | null; maxWidth: number }>(
+        (acc, item) => {
+          const { maxWidth } = acc;
+          if (!("args" in item)) {
+            return acc;
+          }
+          const argsLength =
+            mapInstructionsArgsByType(item.args, NumeralSystem.HEXADECIMAL, 0, program)
+              ?.map((x) => x.value)
+              .join("").length ?? 0;
+
+          if (argsLength > maxWidth) {
+            return { widestItem: item, maxWidth: argsLength };
+          } else {
+            return acc;
+          }
+        },
+        { widestItem: null, maxWidth: 0 },
+      ).widestItem,
+    [programRows, program],
+  );
+
+  if (!widestItem) {
+    return null;
+  }
+
+  return (
+    <InstructionItem
+      index={0}
+      status={Status.OK}
+      isLast={false}
+      onClick={() => {}}
+      instructionMode={instructionMode}
+      programRow={widestItem}
+      currentPc={0}
+      onAddressClick={() => {}}
+      breakpointAddresses={[]}
+      style={{
+        visibility: "hidden",
+      }}
+    />
   );
 }
